% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jscore.R
\name{jscore}
\alias{jscore}
\title{jscore}
\usage{
jscore(truth, pred)
}
\arguments{
\item{truth}{A numeric vector of truth class labels.}

\item{pred}{A numeric vector of predicted class labels.}
}
\value{
Returns the j-score of the clustering assignment.
}
\description{
Calculates the j-score between two clustering assignment.
}
\examples{
truth=c(1,1,1,1,1,2,2,2,2,2,3,3,3,3,3)
pred=c(1,1,1,2,2,2,3,3,3,4,4,4,5,5,5)
jscore(truth, pred)
}
